技能逻辑调整
实现技能自动释放
实现技能直线移动逻辑

+++ Packages/cn.etetet.twsmmo/Scripts/ModelView/Client/Cast/CastComponent.cs
@@ -6,8 +6,8 @@ namespace ET.Client
     /// 释放组件，释放技能
     /// </summary>
     [ComponentOf(typeof(Unit_Client))]
-    public class CastComponent : Entity, IAwake,IDestroy
+    public class CastComponent : Entity, IAwake, IDestroy
     {
-        public Dictionary<long,EntityRef<Cast>> Casts = new Dictionary<long, EntityRef<Cast>>();
+        public Dictionary<int, EntityRef<Cast>> Casts = new Dictionary<int, EntityRef<Cast>>();
     }
 }
